Fix Offline icon not showing everywhere Change default behaviour (don't show spinner when offline)
